+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.20.0] - 2026-08-25
+
 ### Removed
 
 - **BREAKING: `MVPInlineCreateView` and `MVPInlineUpdateView` are gone.** `InlinesMixin` is